pyproject.toml
[project]
name = "tradesagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "valyu",
    "langgraph",
    "langchain-core",
    "python-dotenv"
]
